Complement for ar758x support
Support for the new ar758x platform introduced in ac59008 appeared
to be broken.

This provides some additional fixes to get the target to build.
Most notably, some files are added to provide the pa_adc_Init and
pa_rtc_Init functions in the mdm9x28 swiQmi folder.
Unlike some PA Init functions, these are not part of the general PA
headers but local ones, so local (ie, QMI specific) files are required.

Change-Id: Ie2f252202d3ad0e1b74d1ab6ff66f49cc707502c
Resolves: trac:4713