Embed QMI libraries to prevent dependencies
As to prevent external dependency on QMI libraries, which have led
to tight coupling between the QMI PA and toolchains
(or Yocto meta-swi-bin layer versions), this patch uses static
libraries for everything that is related to QMI headers.

As the QMI headers are already present in proprietary/, this copies the
required .a files from Yocto builds from the different product variants.

Note that pre-built PAs are:
 - not dependent on the .a files anymore: it is not necessary to ship
   them, either in the repository or in the toolchains
 - not shipping .so files from the toolchains anymore: the remaining QMI
   dependencies are on files that change very rarely

Static libraries contained in this patch have been built locally based on
the firmware version mentioned for each target, as firmwares did not contained
patches enabling the generation and deployment of static libraries at the date
of this patch.

Resolves: trac:4929
Change-Id: I17421b17c8c610ba8ae12812288a4840ca76332f